pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "youtube_analysis"
version = "1.0.0"
description = "YouTube Data ETL: extracción, transformación y carga en PostgreSQL (Typer CLI)"
readme = "Readme.md"
requires-python = ">=3.10"
dependencies = [
  "typer[all]>=0.12.1",
  "sqlalchemy>=2.0",
  "psycopg2-binary>=2.9",
  "python-dotenv>=1.0.0",
  "google-api-python-client>=2.100.0",
  "python-dateutil>=2.8.2",
]

[project.scripts]
yt-data = "yt_data.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["yt_data*"]
